moving useAccountDetail to api module
